+++ requirements.txt
@@ -1,2 +1,8 @@
-discord==1.7.3
+# DO NOT include azure-functions-worker in this file
+# The Python Worker is managed by Azure Functions platform
+# Manually managing azure-functions-worker may cause unexpected issues
+
+#azure-functions
+#discord==1.7.3
 requests==2.27.1
+discord-webhook==0.15.0
